[DEV-2694] rename & complete translations
