+++ src/translations/languages/fr.json
@@ -469,6 +469,7 @@
     "Route": "Route",
     "Summary": "Résumé",
     "N/A": "N/A",
+    "delete": "Êtes-vous sûr de vouloir supprimer votre <1>itinéraire de paiement {{type}}</1> avec l’<1>ID {{id}}</1>?",
 
     "Payment details": "Détails du paiement",
     "Your payment details at a glance": "Vos détails de paiement en un coup d'œil",
