+++ src/translations/languages/it.json
@@ -469,6 +469,7 @@
     "Route": "Modalità",
     "Summary": "Riepilogo",
     "N/A": "N/A",
+    "delete": "Sei sicuro di voler eliminare il tuo <1>percorso di pagamento {{type}}</1> con <1>ID {{id}}</1>?",
 
     "Payment details": "Dettagli del pagamento",
     "Your payment details at a glance": "I tuoi dettagli di pagamento a colpo d’occhio",
